removing unneeded comments and functions. Moved removed functions to scripts dir. Changed name of geometry.zoom to general, added center_and_rotate
